+++ pkg/kubectl/BUILD
@@ -10,7 +10,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "autoscale_test.go",
-        "cluster_test.go",
         "clusterrolebinding_test.go",
         "configmap_test.go",
         "delete_test.go",
@@ -35,7 +34,6 @@ go_test(
     ],
     library = ":go_default_library",
     deps = [
-        "//federation/apis/federation/v1beta1:go_default_library",
         "//pkg/api:go_default_library",
         "//pkg/api/testapi:go_default_library",
         "//pkg/api/testing:go_default_library",
@@ -83,7 +81,6 @@ go_library(
         "apply.go",
         "autoscale.go",
         "bash_comp_utils.go",
-        "cluster.go",
         "clusterrolebinding.go",
         "configmap.go",
         "delete.go",
@@ -114,7 +111,6 @@ go_library(
         "versioned_client.go",
     ],
     deps = [
-        "//federation/apis/federation/v1beta1:go_default_library",
         "//pkg/api:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/api/v1/pod:go_default_library",
